fix(daemon): gate deps with actionable error, guard reclaim against PID reuse, add CLI docs, split integration tests
- gaia daemon preflights fastapi/uvicorn/psutil and points to the [ui] extra
  instead of dying on a cryptic subprocess ModuleNotFoundError (base install).
- terminate_instance verifies the pid is still our daemon (cmdline references
  gaia.daemon) before killing, so a reused PID after a hard crash is not killed.
- add a gaia daemon stanza to docs/reference/cli.mdx.
- move the 8 subprocess/HTTP daemon tests to tests/integration/; keep the pure
  instance.json logic tests in tests/unit/.
