+++ src/gaia/cli.py
@@ -6836,6 +6836,36 @@ def handle_agent_import(args):
         sys.exit(1)
 
 
+def _check_daemon_deps():
+    """Fail loud if the daemon's runtime deps (extras-only) are missing.
+
+    ``gaia daemon`` is a base console command but the daemon process needs
+    ``fastapi``/``uvicorn``/``psutil``, which live in the ``[ui]``/``[api]``/
+    ``[dev]`` extras. On a base install the spawned daemon would die on
+    ``ModuleNotFoundError`` and surface a cryptic "process exited early"; name
+    the real cause and the fix instead.
+    """
+    import importlib.util
+
+    missing = [
+        pkg
+        for mod, pkg in (
+            ("fastapi", "fastapi"),
+            ("uvicorn", "uvicorn"),
+            ("psutil", "psutil"),
+        )
+        if importlib.util.find_spec(mod) is None
+    ]
+    if missing:
+        print(
+            "❌ `gaia daemon` needs packages not in the base install: "
+            + ", ".join(missing)
+            + '.\n   Install the daemon extras:  pip install "amd-gaia[ui]"'
+            "  (or [api]/[dev])."
+        )
+        sys.exit(1)
+
+
 def handle_daemon_command(args):
     """Handle `gaia daemon status|stop|restart|logs|start`."""
     action = getattr(args, "daemon_action", None)
@@ -6845,6 +6875,7 @@ def handle_daemon_command(args):
             "available actions (start, status, stop, restart, logs)."
         )
         return
+    _check_daemon_deps()
     if action == "start":
         _handle_daemon_start()
     elif action == "status":
